[iOS] Add IsSupported() to PhotosService
This CL adds a IsSupported() method to PhotosService. The IsAvailable()
method already returns if the service can be used to perform operations
NOW e.g. upload an image. However, another method is needed to determine
whether the service is at all supported i.e. whether the service can be
used to perform operations at some point in the future.

This will be used to determine whether Save to Photos settings can be
shown in Settings, which should be the case even if the service is not
"available" e.g. an operation is already ongoing but should not happen
if the service is not "supported".

Bug: 1478752
Change-Id: Ie8f3b89ac79ca028e436860caa975f90a1132dd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4872779
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1198285}