[sync] Add field in SyncablePrefsMetadata for pref merge behavior
This CL introduces a new field in SyncablePrefsMetadata to specify the
merge behavior for each syncable preference.

Currently there exist methods in PrefModelAssociatorClient which help
check whether a pref is mergeable or not but they are not convenient to
use. The syncable prefs database would later be the source of truth for
this info.

Bug: 1474375
Change-Id: Icfe70884b56eaaa51047512617467800b52e1419
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874056
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1198290}