[iOS] Add DownloadsSettingsTableViewController
This CL adds the DownloadsSettingsTableViewController class which will
be owned by a DownloadsSettingsCoordinator and pushed on top of the
Settings table view.

For now it contains only one section for Save to Photos, with an
IdentityButtonItem to let the user open the Save to Photos account
selection screen, as well as a switch to enable/disable the account
picker step every time the user saves an image to Google Photos.

Bug: 1478752
Change-Id: I143e8510dfb02628b002b6cf78e03f8c63cf1d73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4853506
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198329}