Update the profile when the active user has changed
There is a known issue when the user with OWNER ACCOUNT (the very first ACCOUNT that is logged in) logs in to the Chromebook, the profile is not re-created and they will continue to use the incognito sign-in profile. This causes crashes/misbehaviors for pref service or identity manager.

This issue is not witnessed for users who are NOT the OWNER ACCOUNT, e.g. the second and third account. For those users, their Profile instance is always created after their successful login, and that Profile instance correctly reflects their user data.

This CL fixes that issue by adding an Observer to the Mediator that listens to the "ActiveUserChanged" event. When the OWNER ACCOUNT user logs in, "ActiveUserChanged" is fired because the active user has switched from a null user (incognito profile) into a real OWNER user (real profile).

The method used in this CL is inspired by those applied in:
* Magnifier Manager: https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ash/accessibility/magnification_manager.cc;l=181-188

* System Clock:
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ash/system/system_clock.cc;l=83-90;drc=ea0d7cb33d730a591f481c84881dd8d009a6b948

Bug: b:298865729

Change-Id: Iad2dfa62175e2039d54a28407714c70afd67ff6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4835286
Commit-Queue: Chuong Ho <hdchuong@google.com>
Reviewed-by: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198355}