Remove stale test expectations for view-timeline-range-animation
Previously several tests were running slowly on the Mac when using a
threaded virtual test suite. This test was taking 4 to 5 seconds to
run when it passes, which is too close to the deadline.  Looking at
more current numbers, we are seeing the test take less than 1 second
on average now.  Should be safe to remove the test expectation entries.

Bug: 1467769
Change-Id: I52f6944f0deeacccc422559752fc302d8e34828d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873659
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198376}