[iOS] Use a socket to communicate from run_web_tests.py to a content shell
The iOS web test infrastructure previously employed a file to
temporarily pass a test file name. However, the execution of
web tests was slow due to the extended time it took for reading
and writing to the file. This CL introduces the use of a socket
in place of the file for the web testing infrastructure.

The Fuchsia port had also adopted the socket approach to address
a similar issue. So, the iOS port now shares some code with the
Fuchsia port.

Bug: 1421239
Change-Id: I82adbcc9ba863b42fa9cdb7a9e218e42d6c37c20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4852552
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1198379}