Ensure WCO rect gets send to remote ns view even if set before remote view is created.
Somehow on sonoma it appears that remote_ns_view_ is only initialized
after the initial UpdateWindowControlsOverlay calls, resulting in the
remote nsview not being informed of the correct bounding rect. Cache
the WCO rect in WebContentsViewMac so we can set it properly when
creating the remote ns view.

Bug: 1484330
Change-Id: Ic6ecf73b5aef71c9f49079daac20cc7dba90d068
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4875361
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198437}