Re-land: Start 50/50 trial for std::shared_mutex in StatisticsRecorder.
This is a followup to:
https://chromium-review.googlesource.com/c/chromium/src/+/4851484

That CL implemented the support, but didn't enable it yet.
This CL runs the new impl in a 50/50 client-configured trial.
(It's a client trial because this code needs to run before
base::Feature framework is available).

The results are reported using a synthetic trial and metric
that were added in the previous CL.

If --enable-benchmarking is passed, this CL will use the new
implementation.

Reland of:
https://chromium-review.googlesource.com/c/chromium/src/+/4873417

Patchset 2 shows the diff with the original CL, which is a call
to StatisticsRecorder::FindHistogram() during test_suite init,
so that the lock is initialized before any tests are run.
Verified with a local TSAN build with (confirmed failing before):
  ./out/gn-goma-tsan/leveldb_unittests --gtest_filter="EnvTest.RunMany"

Bug: 1123627, 1484340
Change-Id: I6de7ee5fc8c6e775d3cba60e71c3d041fc2eebb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876974
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198492}