chromeos: Add metadata to crosier FeaturedIntegrationTest
Add the metadata from the Tast version of this crosier test.

Move the AddTestInfo() function to a util file to make it accessible to
more tests. Some InteractiveAshTest run in interactive_ui_tests and not
on VM/DUT, so they do not use ChromeOSIntegrationTestMixin and can't use
AddTestInfo() in its current location.

Bug: b:300649094
Test: chromeos_integration_tests --gtest_filter=*FeaturedIntegrationTest*
Change-Id: Icea51a366f8bf18784f1612d584da3c023a820f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4869552
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198504}