Turn off the highlight when the toggle is clicked
I confirmed with UX that the previously read text should still fade, so the simplest solution is to just have a transparent highlight when the highlight is turned off, since we need the rest of it still.

The reason we use an enum for the highlight granularity rather than a simple boolean is because in future versions we'll have more than just one granularity option with the highlight on.

Bug: 1474951
Change-Id: I1c9fd768b6cd96d5e8344d25c34749c44f28c91d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4869617
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Kristi Saney <kristislee@google.com>
Reviewed-by: Lauren Winston <lwinston@google.com>
Cr-Commit-Position: refs/heads/main@{#1198545}