Rename reportContextualWin to reportAdditionalBidWin
When an additional bid wins an auction, the auction reports this win to
the buyer via a function that has a slightly different name than the one
that would be called to report the win of a bid created by a call to
generateBid(). This is described in more detail at
https://github.com/WICG/turtledove/blob/main/FLEDGE.md#64-reporting-additional-bid-wins.

This change renames this function from reportContextualWin to
reportAdditionalBidWin for consistency, since the term additional bid is
used throughout the auction, including in the header used to transport
it (Ad-Auction-Additional-Bid) and the field on the auction config
(additionalBids).

Change-Id: I8ba001237057d0a2f7e7ded2868750d4af7c7076
Bug: 1464874
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874104
Commit-Queue: Orr Bernstein <orrb@google.com>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198548}