viz: use transform on rounded corner bounds after clipping when ..
... merging masks.

CanMergeMaskFilterInfo checks if merging of a pass with its filter masks is possible.
However, it was doing that incorrectly as it only applied the transform
to the rounded corner bounds ignoring the clip_rect. That resulted in
the rounded corner bounds being outside the clip_rect, which meant the
pass with that mask couldn't be merged.

A proper way to do that is to 1) take the bounds 2) apply the clip 3)
apply the transform.

Bug: 1474855, 1382038

Change-Id: I6c945e9053bd5e7e1ad44c562cae857925a07caf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4860244
Auto-Submit: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Commit-Queue: Peter McNeeley <petermcneeley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198549}