Cellular: Fix suppress text message adoption metric rename
Suppress text messages adoption metrics were added in
https://crrev.com/c/4843828. As part of the review we
renamed some histograms. As part of the rename we missed
the rename for the actual constants and their values and
method names. We rename them so that now they match what we
have in histograms.xml.

Test: chromeos_unittest --gtest_filter=TextMessageProvider*
Test: chromeos_unittest --gtest_filter=NetworkMetadataStore*

Bug: b:294186116
Change-Id: Id82179aafa65be043c3000fa1b54f85799edbd65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4866548
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Commit-Queue: Fahad Mansoor <fahadmansoor@google.com>
Reviewed-by: Gordon Seto <gordonseto@google.com>
Cr-Commit-Position: refs/heads/main@{#1198551}