Attempt to draw promise ring when app is created
When the app list item view is created, force an invoke to
ItemProgressUpdate() to create the progress ring layer and updated if
needed. This is specially necessary for the ring to show on the browser
tests.
Also, modify the token for the ring to paint accordingly in dark or
light mode.

Bug: None
Change-Id: I553dbdf08ade0d4de8176ab6b6fc7b983120864a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877353
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198559}