[QuickStart] Combine QuickStartDecoder methods
Add a new method DecodeQuickStartMessage() to the Mojo interface. This
new method returns a QuickStartMessage, which is a union over all of the
existing decoder types. All of the existing decoder methods are now
implemented using the combined method.

A followup CL will remove the old decoder methods from the Mojo
interface and instead use the DecodeQuickStartMessage() method to
decode all messages. I avoided doing that in this CL in an effort to
keep the line count down and make this easier to review.

Bug: b/294253611
Test: Existing tests pass.
Change-Id: I74ed5e396a6381f0f1cd5e777108898f487f909f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4832468
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Curt Clemens <cclem@google.com>
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Reviewed-by: Brittany Hartmire <bhartmire@google.com>
Cr-Commit-Position: refs/heads/main@{#1198602}