infra: Support links to mirrored/mirroring builders in descriptions
This should allow a viewer to easily map between a CI bot and any try
bot that mirrors it (or vice versa), so long as the builders
involved have migrated their recipe configs to starlark. (We assume
most every chromium bot has.)

And this enables the new descriptions for only builders mirroring
or being mirrored by "GPU FYI Mac Builder". This builder has a lot of
mirrors (see the cr-buildbucket.cfg diff). So we assume it's a good
check of how the descriptions will look. Will expand to all builders
if it looks good.

Note that with the completion of crbug.com/1344925, the description
should be stored on a per-build basis. Meaning: if the description
changes, the description displayed on older builds will still show the
description that was live when the build ran. This is preferred since
it enables us to store somewhat volatile information in the
description. But that may confuse folks if they read the description
from old builds and assume it applies to the builder now. Not sure
how to resolve that. (Maybe show a warning in milo if the old
description != current description?)

Without this change, main.star takes about 3.23s locally. With it, it
takes about 3.28s. (That's the average across 10 runs.)

Bug: 1227778
Change-Id: I50de17c2147669104ebfdc84f7e9ddcd9982a5fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4868655
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1198623}