Set ssp-buffer-size to the default of 8.
The ssp-buffer-size limit is used to decide if stack protector
checking will be added to a function, if a local variable array
of size 8 or greater is present, the function gets the stack
protector prologue and epilogue added.

For reasons old as time, we set it to 4, which triggers
an interesting LLVM bug: https://github.com/llvm/llvm-project/issues/66709

The result of this bug was increased binary size with stack
protector added on functions that do not need it. The compiler
mistakenly counts structure padding as a local array, and adds
the stack protector. With the default of 8 set by this CL,
the unnecessary addition will not normally happen, since the
padding size added to structures will be smaller than 8.

The result will be a decreased binary size (-272K on
android-binary-size).

Change-Id: Ifdd70a975c20a76551c43e6d188c506b8c5913e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874869
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Bob Beck <bbe@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198647}