infra: Roll try-CI mirror builder descriptions to every builder
The description on the following bot doesn't look too bad:
https://ci.chromium.org/ui/p/chromium/builders/ci/GPU%20FYI%20Mac%20Builder

So this rolls it out to every builder. This also removes the "partially"
bit in the description since that might be a little confusing for a
chrome dev unfamiliar with infra mirroring. It was supposed to hint at
when e.g. a child tester on CI is mirrored by a compiling+testing bot
on try. But that might be more details than folk really need to be
aware of.

Bug: 1227778
Change-Id: I0b697f0e5e1edb6b1fa583c4d72b15b3f6219be4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877773
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198679}