Lacros: Fix the crash in preference_api.cc due to invalid pref path.
PreferenceEventRouter::OnAshGetSuccess() is the callback function
for crosapi::mojom::prefs::GetExtensionPrefWithControl(),
which could return absl::nullopt for pref value if it is called
with an invalid pref path. The callback code does not check and
handle this case, therefore, resulting in crash. This cl fixes
the issue.

Bug: 1483928
Change-Id: I319b0c247dd85961fe65ddd859fb06d4bc8d966d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874572
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198740}