snap-group: Fix duplicate group items in overview when minimized
This cl fixed the issue that duplicate overview group items will
show in overview when the windows in a snap group are both minimized.
The logic in `ShouldExcludeForOverview()` should be updated to consider
the scenario when windows are minimized with no default snapped window.

Fixed: b/300703898
Demo: http://b/300703898#comment2
Test: Added unit tests + manual
Change-Id: I64af9338154cfd68476aed497f6227731d67a473
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4870385
Commit-Queue: Michele Fan <michelefan@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198758}