Update local_device_gtest_run.py to support "--force-main-user"
This adds the ability to run applicable adb shell commands as the
main user on the device, Useful when the main user is a secondary user,
e.g. on Android Automotive OS, for gtest type of test suites.

This is a followup of the CL https://crrev.com/c/4803162

Bug: b:293175593
Change-Id: I1636b357cf8391827fca230bfc10f4f3d79602f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874755
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198772}