Fix client window freeze when starting gestures with more than 2 fingers
When a gesture with more than 2 fingers (e.g. 3-finger pinch) starts on
a client window (e.g. Lacros PiP window), it leads to an inconsistent
state and the window becomes immovable until the window is closed.

This commit removes the code that was updating the `gesture_target_` to
`nullptr` in `ToplevelWindowEventHandler::CompleteDrag()`.
`gesture_target_` should not be updated to `nullptr` in `CompleteDrag()`
because:

- the existence of `gesture_target_` is used elsewhere to determine if
  an ongoing event has originated from a touchscreen or a mouse, and

- `CompleteDrag()` is called even when a gesture hasn't completely ended
  (i.e.  when at least one finger is still on the screen, like when a
  more-than-2-finger gesture starts).

If the gesture successfully completes then
`UpdateGestureTarget(nullptr)` is called when the last finger leaves the
screen in `ToplevelWindowEventHandler` anyways, so it should be
unnecessary to call it in `CompleteDrag()`.

Bug: b/301013053
Test: manually verified that the window stays movable after
more-than-2-finger gestures

Change-Id: If4956a0d6846e0b6ddcf2a715bb5624074ed8c1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4876337
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Masa Fujita <massan@google.com>
Auto-Submit: Masa Fujita <massan@google.com>
Cr-Commit-Position: refs/heads/main@{#1198857}