[ios] Fix crash in AutofillImageFetcherImpl::ResolveCardArtImage
The image to resolve can be empty if the server URL served no image
content. This currently causes a CHECK crash when we try to convert
the image back to a UIImage. Fix this crash by checking that the
image is valid before trying to do anything to it.

Bug: 1484797
Change-Id: I6eff33f3326cc2a5166fda11e23736f722e98787
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4878696
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Commit-Queue: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198925}