Fix logic to prevent multiple calls to ApplyAddressFixesAndCleanups for sync-the-transport
ApplyAddressFixesAndCleanups and ApplyCardFixesAndCleanups has multiple
call sites.
[1] From PersonalDataManagerCleaner::CleanupDataAndNotifyPersonalDataObservers
[2] PersonalDataManagerCleaner::SyncStarted

The call from [2] should happen if the user is syncing and have enabled the type in sync-the-feature mode or if the user enabled the relevant user selectable in the sync-the-transport mode.

The call from [1] should happen in all the other cases where the user
is not syncing at all or the user has not enabled the type.

Prior to this CL, there would be calls from both [1] and [2] to the
ApplyAddressFixesAndCleanups in sync-the-transport mode when the user
would have enabled the user selectable type as
`sync_service_->IsSyncFeatureEnabled()` would return false in this case.
This CL ensures that in this case, the call happens from only [2].

Also, the ApplyAddressFixesAndCleanups should happen once both the
AUTOFILL_PROFILE and CONTACT_INFO sync has started as the dedupe logic
uses the contact info profiles for removing the duplicate local profile.

Bug: 1477292
Change-Id: I0f812d70a1a7699618bdc1260fae7998789d2b23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874441
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198945}