Make -webkit-box-direction non-inherited.
This was inherited! Why? No idea. Likely copied from "direction"
originally and not caught. Firefox, EdgeHTML don't have this property
as inherited so likely web compatible.
https://www.software.hixie.ch/utilities/js/live-dom-viewer/?saved=12018

The specification at this stage has it as non-inherited as well:
https://www.w3.org/TR/2009/WD-css3-flexbox-20090723/#displayorder

Behind the Blink flag: NonInheritedWebkitBoxDirection in case someone
was relying on this behaviour.

Change-Id: Ic5b3c1ff741d2e91ea734465b5a2e33e0e598a4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4873849
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1198960}